Backend/Ai/requirements.txt
flask
flask.core
fastapi
uvicorn
scikit-learn
pdfplumber
python-dotenv
google-generativeai
sentence-transformers 
